Extracting database settings from settings.py
Now database settings should be returned by a function called
get_db_settings() which should be put in a file called
database_settings.py that resides in the same folder as settings.py.

Here is the format of get_db_settings() function:

def get_db_settings():
  return {
    'default': {
      'ENGINE': 'django.db.backends.mysql',
      'NAME': '<name here>',
      'USER': '<user name here>',
      'PASSWORD': '<password here>',
      'HOST': '<server name here>',
      'PORT': '<part at servere here, usually 3306 for MySQL>',
    }
  }

Referencing #6 here as well so that we add information about this
to the setup notes.

#5