Adding persons page and moving static and templates files to hadiths project.
The static and template files were moved to hadiths application because
they are specific to that project (at least for now). Additionally,
there shouldn't be any files in the root static directory (which
is assigned to STATIC_ROOT), as that folder's content should be
generated by a call to 'collectstatic' command of manage.py. More info
here:

http://stackoverflow.com/a/12161409/196697

Quoting the answer by Simeon Visser there: "Your STATIC_ROOT directory
should be empty and all static files should be collected into that
directory (i.e., it should not already contain static files)."

#4