Show error messages coming from the server in alerts.
So rather than just saying we failed to serve the request, we also
let the user know what the problem was.

Currently, this helps in showing the correct reason for not being
able to make a change if the user doesn't have permission to make
that change. However, we shouldn't even allow the user to make a
change if he/she doesn't have permission to do so. This will be dealt
with in #38.