Fix migration problem.
This problem happens when we are just creating the database tables from
the migrations or we are adding new permissions. The reason it happens
is beacuse the API views calls the method `Permission.get_code_by_name`
which makes a request to the database to retrieve the required
permission. Now if the permissions table is not yet created or the
required permission hasn't been added yet, Django raises an exception
even when we try to run migrations. This is because for Django to run
migrations, it will have to load the application's Python modules,
one of them is the API views module (apiviews.py) which calls
`Permission.get_code_by_name`.

#49