Add two hadiths initially and remove Person.ref field
It is the same hadith actually, but one version of it from a Shia
book and the other from a Sunni book. The reason for this is to
avoid the project being perceived is biased towards one sector of
Islam.

While working on this, I also removed the reference field, since it
didn't seem to provide much benefit. A big part of the data will
(inshallah) be entered using the website. As for data entered
programmatically, I don't think it is easy to employ the reference
field, as the data would practically contains large number of
persons and creating references to each one of the will probably be
much more difficult than trying to find the persons by other fields
like full name.