feat(offline-dev): Support running the website without internet connection
JavaScript, CSS, and font files were cached under `3rdparty-cache/`
directory and a copy of `index.html` called `index_offline.html` was
created to reference those files instead. A flag was also added to
`settings.py` to allow the developer to specify that he/she wants to
run the website in offline mode.

An offline mode of Facebook service was also created which fake the
original Facebook service.

Finally, authentication in `auth.py` was also changed to automatically
return the first user and thus fake authentication.

#126