feat(chains): Creating custom serializer for chains
The custom serializer is called ChainSerializer. It allows specifying
the persons of a chain as a JSON array and it takes the responsibility
of creating, updating, and deleting the relationship rows between
chains and persons. #56 #57

The serializer also raises a validation error if the same person is
repeated in the persons of a chain. #139