feat(h-t-rel): Add extra information to hadith-tag relation
The extra information are the `added_on` and `added_by` fields. This
was done by creating a standalone table for the relationship between
hadiths and tags and adding the required fields to it. Then,
`HadithSerializer` was updated such that it handles the creation
of relationship records itself and set the `added_by` field by
itself.

Close #74