feat(chains): Display chain tree in hadith page
The tree is built from the list of chains of a certain hadith. The
resulting tree is then displayed using the tree directive.

To be able to efficiently find the name of the persons involved in
the chain tree, a single request with multiple IDs is done. To enable
this to happen, I had to modify the URL of the person resource class
a little bit. It is easier explained in code than words, so check
the changes in `services.ts` in this commit for more information.

#58