feat(caching): Implement CacheableResource and use it in selector directive
CacheableResource is an improvement over caching support in Angular's
ngResource. In the latter, if you make two simultaneous, or very close
to each other, to fetch the same resource, two requests will be made.
CacheableResource makes only one request. This is very useful in case
different parts of the same page requires the same resource. For
example, in the hadith page, requests to the `persons/` API are made
by both the detail of the hadith and the chain tree.