feat(home-page): Change ResourceableCache to support non-number IDs
This was done using generics and now the user of CacheableResource
and Entity needs to specify the type of the ID they want.

While working on this, `getRestfulUrl` was changed such that if only
one ID is requested, it appends the ID to the path (/<id>) rather
than passing the ID as a filter (/?id=<id>).