feat(ex-emails): Enable sending emails to adimns on exceptions
This feature got disabled after we enabled logging because we didn't
use the `mail_admin` handler in the `django.request` logger. I can
see that at some point we should probably remove it again because
we might start receiving unnecessary emails, but for now I guess it
is good to have it to be alerted when a problem happens.

#157