fix(build): Disable running migration to create index during test
While working on #228, a migration was added to create an index for
`text` and `simple_text` fields of `Hadith` model. In addition to the
type of the index being specific to Postgres (GIN index), which
caused the build to fail because a test database doesn't support this
type of index, it is not a good idea to have an index during test.
This commit add an if condition to not run that migration during test
mode.