fix(deploy.sh) Postpone the creation of venv in deploy.sh
We used to create a python virtualenv before copying the app to the
deployment path, then copy everything to the deployment path. This
means the venv was created in path but then moved to another path,
which causes some problems because the venv has the paths (e.g.
package paths) hard coded and -as such- we were not able to, for
example, run `python manage.py` on the web server. This commit fixes
the problem by postponing the creation of the venv to after copying
the app to the deployment path.

Close issue #294