fix(resource) Fix misuse of cache in resource.ts
The function `getByMultipleIdsWithCache` was passing `entity.toString()` to `Cache.put` function instead of `entity.id.toString()`, which caused the `Cache.put` function to fail.

I also added pre-checks to `Cache.get` and `Cache.put` functions so it is obvious in the future if an error is caused by sending a `null`/`undefined` key to those functions.

Close #333