feat(data) Progress towards importing Al-Kafi hadiths
The scripts almost work and tried on the first three volumes. It does seem to be working, though I actually only skimmed the generated JSON instead of going through the hadiths one by one for obvious reasons.

One problem I still need to address is correctly matching kitabs and babs to chapters and sections, respectively. Right now, the 'bab' is mapped to chapter and nothing is mapped to section.

Issue #83